+++ .github/CHANGELOG.md
@@ -2,6 +2,7 @@
 1. [Game] Enhancement - Added Random Events
 2. [Game] Enhancement - Changed uiElements member variable to a dynamic array
 3. [Game] Bug fix - Fix crashing on game over screen on some localizations
+4. [Game] Enhancement - Added text entry for username
 
 # Changes from v1.0.0 to v1.1.0
 1. [Global] Enhancement - Added Changelog